renamed label and helper components; adjusted helper text to newest design
