--- packages/assets/src/scss/_form-label.scss
@@ -1,16 +1,0 @@
-@use 'variables' as *;
-
-.ids-form-label {
-    font-size: $text-font-size-s;
-    color: var(--ids-form-label-default-text-color, #{$color-neutral-160});
-
-    &--required {
-        &::after {
-            content: '*';
-        }
-    }
-
-    &--error {
-        color: var(--ids-form-label-error-text-color, #{$color-error-80});
-    }
-}
