+++ packages/assets/src/scss/styles.scss
@@ -8,5 +8,5 @@
 @use 'fonts';
 @use 'icons';
 @use 'input';
-@use 'form-label';
-@use 'form-helper';
+@use 'label';
+@use 'helper-text';
