--- packages/components/src/FormHelper/index.ts
@@ -1,5 +1,0 @@
-import FormHelper from './FormHelper';
-import { FormHelperProps } from './FormHelper.types';
-
-export default FormHelper;
-export type { FormHelperProps };
