--- packages/components/src/FormLabel/index.ts
@@ -1,5 +1,0 @@
-import FormLabel from './FormLabel';
-import { FormLabelProps } from './FormLabel.types';
-
-export default FormLabel;
-export type { FormLabelProps };
