+++ packages/components/src/HelperText/index.ts
@@ -1,0 +1,5 @@
+import HelperText from './HelperText';
+import { HelperTextProps } from './HelperText.types';
+
+export default HelperText;
+export type { HelperTextProps };
