+++ packages/components/src/Label/index.ts
@@ -1,0 +1,5 @@
+import Label from './Label';
+import { LabelProps } from './Label.types';
+
+export default Label;
+export type { LabelProps };
